Rename object key transformation function and add pascal to camel case function
